♻️ Update `generate_text` function to use `job` module instead of `curl`
The `generate_text` function for the `acommit` Lua module is updated to use the `job` module instead of the `curl` module. This is done to ensure the function is no longer blocking and can be executed asynchronously. A new callback parameter is added to the function with the generated text as a parameter. Job module's `on_exit` function is used to execute the callback function passing generated text as a parameter.

This commit will improve the performance of the function's execution, which previously waited for a response from the `curl` command.